+++ src/main/java/br/com/techchallenge/adapters/dto/Restaurante/RestauranteRequestDTO.java
@@ -1,6 +1,6 @@
 package br.com.techchallenge.adapters.dto.Restaurante;
 
-import br.com.techchallenge.domain.Endereco;
+import br.com.techchallenge.domain.endereco.Endereco;
 
 public record RestauranteRequestDTO(
         String nome,
