+++ src/main/java/br/com/techchallenge/domain/donoRestaurante/DonoRestaurante.java
@@ -1,6 +1,6 @@
 package br.com.techchallenge.domain.donoRestaurante;
 
-import br.com.techchallenge.domain.Endereco;
+import br.com.techchallenge.domain.endereco.Endereco;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
